notification hisory work done, order error mesage shown

+++ app/Http/Controllers/Api/OrderController.php
@@ -95,6 +95,7 @@ class OrderController extends Controller
             }
             else {
                 $cart=Cart::where('user_id',auth()->user()->id)->first();
+                if($cart==null){return ApiHelper::apiResult(false,HttpResponse::HTTP_OK, 'Can not place order because cart is empty');}
                 $cart->delete();
              event(new OrderSuccessEvent($order));
                 return (new OrderResource($order));
